# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Board:
    properties:
      created_at:
        example: "2025-01-15T09:00:00Z"
        type: string
      description:
        example: Planning board for Q3 product initiatives
        type: string
      id:
        example: 2af3e6a1-5c4d-4b7e-8f1a-3d9c6b2e0a85
        type: string
      name:
        example: Q3 Product Roadmap
        type: string
      owner_id:
        example: 7bc4f9d2-1a3e-4c8f-9b2d-5e7a0c4f1d63
        type: string
      updated_at:
        example: "2025-01-15T09:00:00Z"
        type: string
    type: object
...
  domain.CreateBoardInput:
    properties:
      description:
        example: Planning board for Q3 product initiatives
        maxLength: 500
        type: string
      name:
        example: Q3 Product Roadmap
        maxLength: 100
        type: string
    required:
    - name
    type: object
...
  domain.UpdateBoardInput:
    properties:
      description:
        example: Planning board for Q3 product initiatives
        maxLength: 500
        type: string
      name:
        example: Q3 Product Roadmap
        maxLength: 100
        type: string
    required:
    - name
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: error description
        type: string
    type: object
...
host: 204.168.173.88:8080
info:
  contact:
    name: Mykyta Kuzminov
    url: https://github.com/mykytakuzminov
  description: |-
    REST API for managing boards, columns and tasks with JWT authentication.
    Supports token-based auth, rate limiting, pagination, sorting and filtering.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Sprintly API
  version: "1.0"
paths:
...
  /boards:
    get:
      description: Returns all boards owned by the authenticated user. Supports pagination
        and sorting.
      parameters:
      - description: 'Number of results per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of results to skip (default: 0)'
        in: query
        name: offset
        type: integer
      - description: 'Field to sort by: name, created_at, updated_at (default: created_at)'
        in: query
        name: sort
        type: string
      - description: 'Sort direction: ASC or DESC (default: ASC)'
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of boards
          schema:
            items:
              $ref: '#/definitions/domain.Board'
            type: array
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all boards
      tags:
      - boards
    post:
      consumes:
      - application/json
      description: Creates a new board owned by the authenticated user.
      parameters:
      - description: Board data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateBoardInput'
      produces:
      - application/json
      responses:
        "201":
          description: Board created successfully
          schema:
            $ref: '#/definitions/domain.Board'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a board
      tags:
      - boards
...
  /boards/{id}:
    delete:
      description: Deletes a board by ID. Only the board owner can perform this action.
      parameters:
      - description: Board ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Board deleted successfully
        "400":
          description: Invalid board ID format
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Caller is not the board owner
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Board not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a board
      tags:
      - boards
    get:
      description: Returns a specific board by its unique identifier.
      parameters:
      - description: Board ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Board data
          schema:
            $ref: '#/definitions/domain.Board'
        "400":
          description: Invalid board ID format
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Board not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get board by ID
      tags:
      - boards
    patch:
      consumes:
      - application/json
      description: Updates the name and/or description of a board. Only the board
        owner can perform this action.
      parameters:
      - description: Board ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated board data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateBoardInput'
      produces:
      - application/json
      responses:
        "204":
          description: Board updated successfully
        "400":
          description: Invalid board ID format or request body
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Caller is not the board owner
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Board not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a board
      tags:
      - boards
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your Bearer token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"